fix(db): fix dataconfig model import issue
