+++ src/database/useSqlite.ts
@@ -1,9 +1,9 @@
-import { createConnection, EntityTarget, getConnectionOptions } from 'typeorm'
+import { createConnection, EntityTarget } from 'typeorm'
+import ORMConfig from './database.config'
 
 export default function useSqlite() {
   const insert = async <T>(data: T, model: EntityTarget<T>): Promise<T> => {
-    const config = await getConnectionOptions('MQTTX')
-    const db = await createConnection(config)
+    const db = await createConnection(ORMConfig)
     const repository = db.getRepository(model)
     return repository.save(data)
   }
